Simplify webpacker config and split up apollo
