+++ app/javascript/components/single_bar_view/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
 import { useQuery } from '@apollo/client';
-import { GET_SINGLE_BAR } from '../../utilities/graphql_utilities';
+import { GET_SINGLE_BAR } from '../../utilities/graphql_tags';
 import SingleBar from '../single_bar';
 import LoadingManager from '../loading_manager';
