+++ config/webpacker.yml
@@ -57,10 +57,11 @@ development:
   # Reference: https://webpack.js.org/configuration/dev-server/
   dev_server:
     https: false
-    host: localhost
+    host: 0.0.0.0
     port: 3035
-    public: localhost:3035
-    hmr: false
+    public: 0.0.0.0:3035
+    hmr: true
+    hot: true
     # Inline should be set to true if using HMR
     inline: true
     overlay: true
